# Separate out query string encoding for non-encoded strings.