--- a/codec-http/src/test/java/io/netty/handler/codec/http/QueryStringEncoderTest.java
+++ b/codec-http/src/test/java/io/netty/handler/codec/http/QueryStringEncoderTest.java
@@ -15,11 +15,11 @@
  */
 package io.netty.handler.codec.http;
 
-import java.net.URI;
-import java.nio.charset.Charset;
-
 import org.junit.Assert;
 import org.junit.Test;
+
+import java.net.URI;
+import java.nio.charset.Charset;
 
 public class QueryStringEncoderTest {
 
@@ -36,6 +36,11 @@ public class QueryStringEncoderTest {
         e.addParam("a", "\u00A5");
         Assert.assertEquals("/foo/\u00A5?a=%C2%A5", e.toString());
         Assert.assertEquals(new URI("/foo/\u00A5?a=%C2%A5"), e.toUri());
+
+        e = new QueryStringEncoder("/foo/\u00A5");
+        e.addParam("a", "abc\u00A5");
+        Assert.assertEquals("/foo/\u00A5?a=abc%C2%A5", e.toString());
+        Assert.assertEquals(new URI("/foo/\u00A5?a=abc%C2%A5"), e.toUri());
 
         e = new QueryStringEncoder("/foo");
         e.addParam("a", "1");