@normanmaurer 2019-12-19
@netty-bot test this please