@normanmaurer 2019-12-19
@franz1981 any other concerns or happy to have this merged ?